requirements.txt
# --- Core Frameworks ---
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1

# --- AI & LLM ---
ollama==0.1.6

# --- Image & PDF Processing ---
pytesseract==0.3.10
opencv-python==4.9.0.80
pdf2image==1.17.0
Pillow==10.2.0
numpy==1.26.4

# --- Utilities ---
python-dotenv==1.0.1
amadeus==9.0.0
python-docx==1.1.0
chromadb==0.4.22
posthog<3.0.0
requests==2.31.0
beautifulsoup4==4.12.3
google-search-results==2.4.2